requirements.txt
web3
eth-account
python-dotenv
eth-brownie
ccxt